feat: define entrypoint
Migrate all the code in codeexplorer.py into a `main()` function,
which involved some refactoring of parts of the code that assumed
script level variables (eg, the adapters all assumed a global `prompt`).

Some data, like tool definitions, is left at module level. Later we
might move that into a `tools.py` file, so I figured it'd be easier
to do that if it was left as-is. Because it's only data, it works just
fine.

Add an entrypoint to pyproject.toml, and update the README.md.