refactor: pass in model to adapter contructor
This is mostly important for the watsonx adapter. Prior to this commit,
it created a ModelInference object for every chat call. This wasn't
very efficient, because the object's init makes a call to watsonx to
validate the model name. By passing the model to the adapter init,
we can create the ModelInference just once during init, saving time
later.