Use git ls-files for file listing when in git repo
Issue: https://github.com/mikerhodes/ai-codeexplorer/issues/2

When in a git repository, the output of `git ls-files`, ie, the files
known to git, is likely to be a _far_ better and more concise listing
of the source code files in the repository than a straight directory
listing (which might include temporary files, caches like __pycache__,
virtualenvs and so on). So it's likely to be much more what a code
exploring AI wants to see!

In particular, folders like venv or node_modules tend to be huge, and
will quickly fill up context windows.